App finished
Added detailsEmployee
Added validations on:
addEmployee. updateEmployee
Used swal api to create a confirm delete button
App finished

+++ employee-managment-frontend/src/app/employee-list/employee-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Employee } from '../employee';
 import { EmployeeService } from '../employee.service';
 import { Router } from '@angular/router';
+import swal from 'sweetalert2';
 
 @Component({
   selector: 'app-employee-list',
@@ -29,14 +30,27 @@ export class EmployeeListComponent implements OnInit{
   }
 
   deleteEmployee(id:number){
-    this.employeeService.deleteEmployee(id).subscribe(data => {
-      console.log(data);
-      this.getEmployees();
+    swal.fire({
+      title : 'Are you sure?',
+      text : 'Confirm if you want to delete this employee',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: "Confirm",
+      cancelButtonText: "Cancel",
+      buttonsStyling: true
+    }).then((result) => {
+      if(result.value){
+        this.employeeService.deleteEmployee(id).subscribe(data => {
+          console.log(data);
+          this.getEmployees();
+          swal.fire('Employee deleted', "Employee has been deleted", 'success')
+        })
+      }
     })
   }
 
   seeDetails(id:number){
     this.router.navigate(['employee-details', id]);
   }
-
 }
